upd refactoring

+++ Assets/Scripts/UI/SetDifficultyDropdown.cs
@@ -7,14 +7,13 @@ namespace UI
     public class SetDifficultyDropdown : MonoBehaviour
     {
         [SerializeField] private GameData gameData;
-        
+
         private TMP_Dropdown _dropdown;
 
         private void Start()
         {
             _dropdown = GetComponent<TMP_Dropdown>();
-            _dropdown.value = (int)gameData!.difficulty;
+            _dropdown.value = gameData!.difficulty;
         }
-
     }
 }
